Add VAT service API client and tests
This commit includes a new client for the Greek VAT service API written in Go. It includes capabilities for retrieving the version of the service used and fetching VAT information. Tests for these functions are also added. The client will automatically load credentials from certain environment variables if they are not supplied. The tests use the assert package from "github.com/stretchr/testify" to check expected results and "github.com/davecgh/go-spew/spew" for pretty printing.

The change allows this software to interact with the Greek VAT service API. This can be valuable in contexts where there is a need to check VAT validity or retrieve related information programmatically, such as building online platforms dealing with invoices or accounting-related tasks.