Add new ConnectionClient interface in clientInterface.go
A new interface, ConnectionClient, has been added in the mydata package to define a set of methods that must be implemented by any type that claims to satisfy this interface. These methods include sending and requesting invoices, requesting income and expense data, and other related actions. This improvement allows other parts of the software to depend on the interface, not directly on the implementing type, facilitating future extensions and modifications.