"Added VAT calculation function and fixed a comment format"
This commit introduces a new function CalculateVAT to the InvoiceVATCategory in the enums.go file. This function helps to calculate the variable VAT rate for different invoice categories. The user can now pass in the amount to get the VAT value, improving versatility and ease of use.

Also, a comment format was fixed for consistency in 'E3_318' ExpenseClassificationTypeStringType.